Move add_report to a helper
