+++ project/tests/test_reports.py
@@ -4,12 +4,7 @@ from project.tests.base import BaseTestCase
 from project.api.models import Report
 from datetime import datetime
 from project import db
-
-def add_report(company_id, data_from, data_to, total_cost, total_tax_cost):
-    report = Report(company_id, data_from, data_to, total_cost, total_tax_cost)
-    db.session.add(report)
-    db.session.commit()
-    return report
+from project.tests.utils import add_report
 
 class TestReportService(BaseTestCase):
